--- src/constants/PageOptions.ts
@@ -1,10 +1,0 @@
-import Option from '../types/Options';
-
-const pageOptions: Option[] = [
-  { value: '4', label: '4' },
-  { value: '8', label: '8' },
-  { value: '16', label: '16' },
-  { value: 'all', label: 'All' },
-];
-
-export default pageOptions;
