--- src/constants/sortOptions.ts
@@ -1,9 +1,0 @@
-import Option from '../types/Options';
-
-const sortOptions: Option[] = [
-  { value: 'newest', label: 'Newest' },
-  { value: 'alphabetically', label: 'Alphabetically' },
-  { value: 'cheapest', label: 'Cheapest' },
-];
-
-export default sortOptions;
