--- src/index.d.ts
@@ -1,7 +1,0 @@
-declare module '*.css' {
-  interface IClassNames {
-    [className: string]: string;
-  }
-  const classNames: IClassNames;
-  export = classNames;
-}
