+++ src/modules/HomePage/components/HotPrices/HotPrices.tsx
@@ -2,13 +2,13 @@ import 'swiper/css';
 import 'swiper/css/navigation';
 import 'swiper/css/pagination';
 
-import { useEffect, useState } from 'react';
-
 import {
   getAccessories,
   getPhones,
   getTablets,
 } from '../../../../api/getProduct';
+import { useEffect, useState } from 'react';
+
 import Product from '../../../../types/Product';
 import SliderProducts from '../../../shared/SliderProducts/SliderProducts';
 
@@ -18,7 +18,6 @@ const HotPrices = () => {
   const [accessories, setAccessories] = useState<Product[]>([]);
   const [totalPages, setTotalPages] = useState(1);
   const [isLoading, setIsLoading] = useState(true);
-  const [isLoadingProduct, setIsLoadingProduct] = useState(true);
   const [isError, setIsError] = useState(false);
 
   const fetchData = async (
@@ -42,19 +41,13 @@ const HotPrices = () => {
     } catch (error) {
       setIsError(true);
     } finally {
-      setTimeout(() => {
-        setIsLoading(false);
-      }, 100);
-      setTimeout(() => {
-        setIsLoadingProduct(false);
-      }, 500);
+      setIsLoading(false);
     }
 
     return 0;
   };
 
   useEffect(() => {
-    setIsLoadingProduct(true);
     setIsLoading(true);
 
     const fetchAllData = async () => {
@@ -78,7 +71,6 @@ const HotPrices = () => {
       products={allProducts}
       totalPages={totalPages}
       isLoading={isLoading}
-      isLoadingProduct={isLoadingProduct}
       isError={isError}
     />
   );
