+++ src/modules/ProductPage/ProductPage.module.css
@@ -1,0 +1,40 @@
+.content {
+  padding: 24px 0 80px;
+}
+
+.title {
+  margin-bottom: 8px;
+}
+
+.quantity {
+  color: var(--font-color-secondary);
+  margin-bottom: 40px;
+  font-weight: 600;
+}
+
+.filter {
+  display: flex;
+  align-items: center;
+  gap: 16px;
+  margin-bottom: 24px;
+}
+
+.label {
+  color: var(--font-color-secondary);
+  font-size: 12px;
+  margin-bottom: 4px;
+  font-weight: 700;
+}
+
+.error {
+  padding: 25px 0;
+  font-size: 20px;
+  color: var(--color-red);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  gap: 20px;
+  height: 50vh;
+  text-align: center;
+}
