+++ src/routes/Root.tsx
@@ -1,0 +1,6 @@
+import { RouterProvider } from 'react-router-dom';
+import { router } from './router';
+
+const Root = () => <RouterProvider router={router} />;
+
+export default Root;
