+++ src/utils/searchHelper.ts
